Changed unsync to remove synced role and moved InfoCommand to par namespace
