+++ src/main/java/com/everneth/emi/commands/par/InfoCommand.java
@@ -1,4 +1,4 @@
-package com.everneth.emi.commands;
+package com.everneth.emi.commands.par;
 
 import co.aikar.commands.BaseCommand;
 import co.aikar.commands.annotation.CommandAlias;
